fix: keep join order
When joining in other tables and sorting on one of the joined rows. Then Knorm would change the order of the output array, so it's locally sorted on the id or other unique property. This happened because an object was used, so because the keys would be numbers ouput of Object.values would be sorted on the key.